[UPD] split views from utils.base into their own cog
