+++ cogs/moderation/reformation.py
@@ -1,7 +1,7 @@
 import nextcord
 from nextcord.ext import commands
 
-from utils.base import ConfirmView, SersiEmbed, ban
+from utils.base import ban
 from utils.cases import get_reformation_next_case_number
 from utils.channels import make_transcript
 from utils.config import Configuration
@@ -9,6 +9,8 @@ from utils.database import db_session, BanCase, ReformationCase, VoteDetails, Vo
 from utils.offences import fetch_offences_by_partial_name
 from utils.perms import permcheck, is_mod, is_senior_mod
 from utils.roles import parse_roles
+from utils.sersi_embed import SersiEmbed
+from utils.views import ConfirmView
 from utils.voting import VoteView, vote_planned_end
